Refactor into connection module
