+++ proxy/core/connection/__init__.py
@@ -1,0 +1,21 @@
+# -*- coding: utf-8 -*-
+"""
+    proxy.py
+    ~~~~~~~~
+    ⚡⚡⚡ Fast, Lightweight, Pluggable, TLS interception capable proxy server focused on
+    Network monitoring, controls & Application development, testing, debugging.
+
+    :copyright: (c) 2013-present by Abhinav Singh and contributors.
+    :license: BSD, see LICENSE for more details.
+"""
+from .connection import TcpConnection, TcpConnectionUninitializedException, tcpConnectionTypes
+from .client import TcpClientConnection
+from .server import TcpServerConnection
+
+__all__ = [
+    'TcpConnection',
+    'TcpConnectionUninitializedException',
+    'TcpServerConnection',
+    'TcpClientConnection',
+    'tcpConnectionTypes',
+]
