+++ tests/core/test_connection.py
@@ -73,7 +73,7 @@ class TestTcpConnection(unittest.TestCase):
         mock_socket.return_value.connect.assert_called_with(
             (str(DEFAULT_IPV6_HOSTNAME), DEFAULT_PORT, 0, 0))
 
-    @mock.patch('proxy.core.connection.new_socket_connection')
+    @mock.patch('proxy.core.connection.server.new_socket_connection')
     def testTcpServerIgnoresDoubleConnectSilently(
             self,
             mock_new_socket_connection: mock.Mock) -> None:
@@ -93,7 +93,7 @@ class TestTcpConnection(unittest.TestCase):
         mock_socket.return_value.connect.assert_called_with(
             (str(DEFAULT_IPV4_HOSTNAME), DEFAULT_PORT))
 
-    @mock.patch('proxy.core.connection.new_socket_connection')
+    @mock.patch('proxy.core.connection.server.new_socket_connection')
     def testTcpServerConnectionProperty(
             self,
             mock_new_socket_connection: mock.Mock) -> None:
